# api/swagger.yml
...
consumes:
    - application/json
produces:
    - application/json
swagger: "2.0"
info:
    description: API documentation
    title: github.com/JehadAbdulwafi/rustion
    version: 0.1.0
paths:
    /api/v1/apps:
        get:
            description: Return a list of App
            operationId: GetAppListRoute
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/appList'
            security:
                - Bearer: []
            summary: Return a list of App
            tags:
                - App
        post:
            description: Update the given App
            operationId: PostAppRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/appPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/app'
            security:
                - Bearer: []
            summary: Update the given App
            tags:
                - App
    /api/v1/apps/{id}:
        delete:
            description: Delete App with ID
            operationId: DeleteAppRoute
            parameters:
                - description: ID of App
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "204":
                    description: Success
            security:
                - Bearer: []
            summary: Delete App with ID
            tags:
                - App
        get:
            description: Return App with ID
            operationId: GetAppRoute
            parameters:
                - description: ID of App
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/app'
            security:
                - Bearer: []
            summary: Return App with ID
            tags:
                - App
        put:
            description: Update the given App
            operationId: PutAppRoute
            parameters:
                - description: ID of App
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/appConfigPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/messageResponse'
            security:
                - Bearer: []
            summary: Update the given App
            tags:
                - App
    /api/v1/articles:
        get:
            description: Retrieves all articles
            operationId: GetAllArticles
            responses:
                "200":
                    description: List of articles
                    schema:
                        $ref: '#/definitions/getArticlesResponse'
            summary: Gets all articles
            tags:
                - articles
        post:
            description: Creates a article
            operationId: PostCreateArticle
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/createArticleRequest'
            responses:
                "200":
                    description: CreateArticleResponse
                    schema:
                        $ref: '#/definitions/messageResponse'
            summary: Creates a article
            tags:
                - articles
    /api/v1/articles/{id}:
        delete:
            description: Deletes a article by ID
            operationId: DeleteArticle
            parameters:
                - description: ID of the article
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: DeleteArticleResponse
                    schema:
                        $ref: '#/definitions/messageResponse'
            summary: Deletes a article
            tags:
                - articles
        get:
            description: Retrieves a article by ID
            operationId: GetArticle
            parameters:
                - description: ID of the article
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: GetArticleResponse
                    schema:
                        $ref: '#/definitions/getArticleResponse'
            summary: Gets a article
            tags:
                - articles
        put:
            description: Updates a article by ID
            operationId: UpdateArticle
            parameters:
                - description: ID of the article
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/updateArticleRequest'
            responses:
                "200":
                    description: UpdateArticleResponse
                    schema:
                        $ref: '#/definitions/messageResponse'
            summary: Updates a article
            tags:
                - articles
...
    /api/v1/auth/forgot-password:
        post:
            description: |-
                Initiates a password reset for a local user, sending an email with a password
                reset link to the provided email address if a user account exists. Will always
                succeed, even if no user was found in order to prevent user enumeration
            operationId: PostForgotPasswordRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/postForgotPasswordPayload'
            responses:
                "204":
                    description: Success
                "400":
                    description: PublicHTTPValidationError
                    schema:
                        $ref: '#/definitions/publicHttpValidationError'
            summary: Initiate password reset for local user
            tags:
                - auth
    /api/v1/auth/forgot-password/complete:
        post:
            description: |-
                Completes a password reset for a local user, using the password reset token sent via email
                to confirm user access, setting the new password if successful. All current access and refresh
                tokens are invalidated and a new set of auth tokens is returned
            operationId: PostForgotPasswordCompleteRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/postForgotPasswordCompletePayload'
            responses:
                "200":
                    description: PostLoginResponse
                    schema:
                        $ref: '#/definitions/postLoginResponse'
                "400":
                    description: PublicHTTPValidationError, type `INVALID_PASSWORD`
                    schema:
                        $ref: '#/definitions/publicHttpValidationError'
                "403":
                    description: PublicHTTPError, type `USER_DEACTIVATED`/`NOT_LOCAL_USER`
                    schema:
                        $ref: '#/definitions/publicHttpError'
                "404":
                    description: PublicHTTPError, type `TOKEN_NOT_FOUND`
                    schema:
                        $ref: '#/definitions/publicHttpError'
                "409":
                    description: PublicHTTPError, type `TOKEN_EXPIRED`
                    schema:
                        $ref: '#/definitions/publicHttpError'
            summary: Completes password reset for local user
            tags:
                - auth
    /api/v1/auth/login:
        post:
            description: Returns an access and refresh token on successful authentication
            operationId: PostLoginRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/postLoginPayload'
            responses:
                "200":
                    description: PostLoginResponse
                    schema:
                        $ref: '#/definitions/postLoginResponse'
                "400":
                    description: PublicHTTPValidationError
                    schema:
                        $ref: '#/definitions/publicHttpValidationError'
                "401":
                    description: PublicHTTPError
                    schema:
                        $ref: '#/definitions/publicHttpError'
                "403":
                    description: PublicHTTPError, type `USER_DEACTIVATED`
                    schema:
                        $ref: '#/definitions/publicHttpError'
            summary: Login with local user
            tags:
                - auth
    /api/v1/auth/refresh:
        post:
            description: Returns a new access token on successful refresh
            operationId: PostRefreshRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/postRefreshPayload'
            responses:
                "200":
                    description: PostLoginResponse
                    schema:
                        $ref: '#/definitions/postLoginResponse'
                "401":
                    description: PublicHTTPError
                    schema:
                        $ref: '#/definitions/publicHttpError'
            security:
                - Bearer: []
            summary: Refresh access token
            tags:
                - auth
    /api/v1/auth/register:
        post:
            description: Returns an access and refresh token on successful registration
            operationId: PostRegisterRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/postRegisterPayload'
            responses:
                "200":
                    description: PostLoginResponse
                    schema:
                        $ref: '#/definitions/postLoginResponse'
                "400":
                    description: PublicHTTPValidationError, type `INVALID_PASSWORD`
                    schema:
                        $ref: '#/definitions/publicHttpValidationError'
                "409":
                    description: PublicHTTPError, type `USER_ALREADY_EXISTS`
                    schema:
                        $ref: '#/definitions/publicHttpError'
            summary: Registers a local user
            tags:
                - auth
...
    /api/v1/channels:
        get:
            description: Return a list of Channel
            operationId: GetChannelListRoute
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/channelList'
            security:
                - Bearer: []
            summary: Return a list of Channel
            tags:
                - Channel
        post:
            description: Update the given Channel
            operationId: PostChannelRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/channelPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/channel'
            security:
                - Bearer: []
            summary: Update the given Channel
            tags:
                - Channel
...
    /api/v1/faqs:
        get:
            description: Return a list of Faq
            operationId: GetFaqListRoute
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/faqList'
            security:
                - Bearer: []
            summary: Return a list of Faq
            tags:
                - Faq
        post:
            description: Update the given Faq
            operationId: PostFaqRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/faqPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/faq'
            security:
                - Bearer: []
            summary: Update the given Faq
            tags:
                - Faq
    /api/v1/faqs/{id}:
        delete:
            description: Delete Faq with ID
            operationId: DeleteFaqRoute
            parameters:
                - description: ID of Faq
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "204":
                    description: Success
                    schema:
                        $ref: '#/definitions/messageResponse'
            security:
                - Bearer: []
            summary: Delete Faq with ID
            tags:
                - Faq
        get:
            description: Return Faq with ID
            operationId: GetFaqRoute
            parameters:
                - description: ID of Faq
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/faq'
            security:
                - Bearer: []
            summary: Return Faq with ID
            tags:
                - Faq
        put:
            description: Update the given Faq
            operationId: PutFaqRoute
            parameters:
                - description: ID of Faq
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/faqPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/messageResponse'
            security:
                - Bearer: []
            summary: Update the given Faq
            tags:
                - Faq
    /api/v1/featured-sections:
        get:
            description: Retrieves all featured sections with their associated articles
            operationId: GetAllFeaturedSectionsWithArticles
            responses:
                "200":
                    description: List of featured sections with articles
                    schema:
                        $ref: '#/definitions/getFeaturedSectionsResponse'
            summary: Gets all featured sections with articles
            tags:
                - featured_sections
        post:
            description: Creates a featured section
            operationId: PostCreateFeaturedSection
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/createFeaturedSectionRequest'
            responses:
                "200":
                    description: CreateFeaturedSectionResponse
                    schema:
                        $ref: '#/definitions/messageResponse'
            summary: Creates a featured section
            tags:
                - featured_sections
...
    /api/v1/featured-sections/{id}:
        delete:
            description: Deletes a featured section by ID
            operationId: DeleteFeaturedSection
            parameters:
                - description: ID of the featured section
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: DeleteFeaturedSectionResponse
                    schema:
                        $ref: '#/definitions/messageResponse'
            summary: Deletes a featured section
            tags:
                - featured_sections
        get:
            description: Retrieves a featured section by ID
            operationId: GetFeaturedSection
            parameters:
                - description: ID of the featured section
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: GetFeaturedSectionResponse
                    schema:
                        $ref: '#/definitions/getFeaturedSectionResponse'
            summary: Gets a featured section
            tags:
                - featured_sections
        put:
            description: Updates a featured section by ID
            operationId: UpdateFeaturedSection
            parameters:
                - description: ID of the featured section
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/updateFeaturedSectionRequest'
            responses:
                "200":
                    description: UpdateFeaturedSectionResponse
                    schema:
                        $ref: '#/definitions/messageResponse'
            summary: Updates a featured section
            tags:
                - featured_sections
    /api/v1/featured-sections/{id}/articles:
        get:
            description: Retrieves a featured section with its associated articles
            operationId: GetFeaturedSectionWithArticles
            parameters:
                - description: ID of the featured section
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: GetFeaturedSectionWithArticlesResponse
                    schema:
                        $ref: '#/definitions/getFeaturedSectionWithArticlesResponse'
            summary: Gets a featured section with articles
            tags:
                - featured_sections
        put:
            description: Updates a featured articles by ID
            operationId: UpdateFeaturedArticles
            parameters:
                - description: ID of the featured section
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/updateFeaturedArticlesRequest'
            responses:
                "200":
                    description: UpdateFeaturedSectionResponse
                    schema:
                        $ref: '#/definitions/messageResponse'
            summary: Updates a featured articles
            tags:
                - featured_sections
    /api/v1/feedbacks:
        get:
            description: Return a list of Feedback
            operationId: GetFeedbackListRoute
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/feedbackList'
            security:
                - Bearer: []
            summary: Return a list of Feedback
            tags:
                - Feedback
        post:
            description: Update the given Feedback
            operationId: PostFeedbackRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/feedbackPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/messageResponse'
            security:
                - Bearer: []
            summary: Update the given Feedback
            tags:
                - Feedback
    /api/v1/feedbacks/{id}:
        delete:
            description: Delete Feedback with ID
            operationId: DeleteFeedbackRoute
            parameters:
                - description: ID of Feedback
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "204":
                    description: Success
            security:
                - Bearer: []
            summary: Delete Feedback with ID
            tags:
                - Feedback
        get:
            description: Return Feedback with ID
            operationId: GetFeedbackRoute
            parameters:
                - description: ID of Feedback
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/feedback'
            security:
                - Bearer: []
            summary: Return Feedback with ID
            tags:
                - Feedback
        put:
            description: Update the given Feedback
            operationId: PutFeedbackRoute
            parameters:
                - description: ID of Feedback
                  format: uuid4
                  in: path
                  name: id
                  required: true
                  type: string
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/feedbackPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/messageResponse'
            security:
                - Bearer: []
            summary: Update the given Feedback
            tags:
                - Feedback
...
    /api/v1/push/test:
        get:
            deprecated: true
            description: Sends a test push message to the current user
            operationId: GetPushTestRoute
            responses:
                "200":
                    description: OK
            security:
                - Bearer: []
            summary: Send test push
            tags:
                - test
    /api/v1/push/token:
        put:
            description: |-
                Adds a push token for the given provider to the current user.
                If the oldToken is present it will be deleted.
                Currently only the provider 'fcm' is supported.
            operationId: PostUpdatePushTokenRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/postUpdatePushTokenPayload'
            responses:
                "200":
                    description: OK
                "404":
                    description: PublicHTTPError, type `OLD_PUSH_TOKEN_NOT_FOUND`
                    schema:
                        $ref: '#/definitions/publicHttpError'
                "409":
                    description: PublicHTTPError, type `PUSH_TOKEN_ALREADY_EXISTS`
                    schema:
                        $ref: '#/definitions/publicHttpError'
            security:
                - Bearer: []
            summary: Adds a push token to the user
            tags:
                - push
    /api/v1/streams:
        get:
            description: Return a list of Stream
            operationId: GetStreamListRoute
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/streamList'
            security:
                - Bearer: []
            summary: Return a list of Stream
            tags:
                - Stream
        post:
            description: Update the given Stream
            operationId: PostStreamRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/createStreamPayload'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/stream'
            security:
                - Bearer: []
            summary: Update the given Stream
            tags:
                - Stream
    /api/v1/streams/events:
        post:
            description: ' Stream event'
            operationId: PostStreamEventRoute
            parameters:
                - in: body
                  name: Payload
                  schema:
                    $ref: '#/definitions/streamEvent'
            responses:
                "200":
                    description: Success
                    schema:
                        $ref: '#/definitions/streamEventResponse'
            summary: Stream event
            tags:
                - Stream
...
definitions:
    app:
        properties:
            config:
                type: string
            createdAt:
                type: string
            id:
                format: uuid4
                type: string
            name:
                type: string
            updatedAt:
                type: string
            userID:
                format: uuid4
                type: string
        required:
            - id
            - userID
            - name
            - config
        type: object
    appConfigPayload:
        properties:
            config:
                type: string
        required:
            - config
        type: object
    appList:
        items:
            $ref: '#/definitions/app'
        type: array
    appPayload:
        properties:
            config:
                type: string
            name:
                type: string
        required:
            - config
            - name
        type: object
...
    faqList:
        items:
            $ref: '#/definitions/faq'
        type: array
...
    getArticleResponse:
        properties:
            article:
                $ref: '#/definitions/article'
        type: object
    getArticlesResponse:
        items:
            $ref: '#/definitions/article'
        type: array
...
    messageResponse:
        properties:
            message:
                example: created successfully.
                type: string
        type: object
...
securityDefinitions:
    Bearer:
        description: |-
            Access token for application access, **must** include "Bearer " prefix.
            Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
        in: header
        name: Authorization
        type: apiKey
        x-keyPrefix: 'Bearer '